requirements.txt
torch>=2.1
torchvision>=0.16
transformers>=4.41
sentence-transformers>=2.6
pillow>=10.0
numpy>=1.24
scikit-learn>=1.3
tqdm>=4.66
networkx>=3.2
